feat: Retrospective 관련 코드에 문서 주석 작성
