+++ Retrospective/Core/Resources/AppColors.swift
@@ -10,9 +10,6 @@ import SwiftUI
 extension ShapeStyle where Self == Color {
 
     /// 주요 글자색
-    ///
-    /// * 라이트 모드: 검정색
-    /// * 다크 모드: 흰색
     static var label: Color {
         .init(UIColor.label)
     }
